python_api/requirements.txt
openai
flask
requests
pdf2image
# apt-get install poppler-utils

python-dotenv
